SDL_timer.c: silence a minor warning.

+++ src/timer/SDL_timer.c
@@ -539,7 +539,7 @@ void SDL_InitTicks(void)
                         SDL_TimerResolutionChanged, NULL);
 
     tick_freq = SDL_GetPerformanceFrequency();
-    SDL_assert(tick_freq > 0 && tick_freq <= SDL_MAX_UINT32);
+    SDL_assert(tick_freq > 0 && tick_freq <= (Uint64)SDL_MAX_UINT32);
 
     gcd = CalculateGCD(SDL_NS_PER_SECOND, (Uint32)tick_freq);
     tick_numerator_ns = (SDL_NS_PER_SECOND / gcd);
